# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem manajemen rental PlayStation Byone Arena. Mendukung realtime via WebSocket untuk aplikasi mobile dan TV Android.",
        "title": "Byone Arena API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Byone Arena Support",
            "email": "support@byone-arena.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/sessions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengembalikan daftar seluruh sesi rental (aktif, selesai, dibatalkan)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sesi Rental"
                ],
                "summary": "Ambil semua sesi rental",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.Session"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/active": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengembalikan sesi yang sedang berjalan saat ini di semua konsol",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sesi Rental"
                ],
                "summary": "Ambil semua sesi aktif",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.Session"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/start": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Memulai sesi rental untuk konsol dan langsung menyelesaikan pembayaran di depan dalam satu transaksi.\\n\\n**Alur:**\\n1. Konsol harus berstatus `available`\\n2. Harga dihitung dari `bookedDurationMinutes × pricePerHour / 60`\\n3. Diskon otomatis (happy hour, member) diterapkan\\n4. Voucher opsional diterapkan jika `voucherCode` diberikan\\n5. `cashReceived` harus ≥ harga setelah diskon\\n\\nResponse mencakup `session` (dengan `endScheduledAt` untuk countdown) dan `payment` (dengan `changeAmount` kembalian).\\n\\nEvent realtime `session_started` dikirim ke semua client WebSocket.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sesi Rental"
                ],
                "summary": "Mulai sesi rental + pembayaran di depan",
                "parameters": [
                    {
                        "description": "Data sesi dan pembayaran",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/usecase.StartSessionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usecase.StartSessionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Konsol tidak tersedia, uang kurang, atau voucher tidak valid",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengembalikan detail satu sesi rental beserta data konsol dan pelanggan",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sesi Rental"
                ],
                "summary": "Ambil sesi berdasarkan ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.Session"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/{id}/cancel": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membatalkan sesi rental aktif tanpa tagihan. Konsol dikembalikan ke status available. Event realtime dikirim.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sesi Rental"
                ],
                "summary": "Batalkan sesi rental",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Sesi tidak aktif atau tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/sessions/{id}/end": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengakhiri sesi rental aktif, menghitung durasi dan total harga, membebaskan konsol. Event realtime dikirim.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sesi Rental"
                ],
                "summary": "Akhiri sesi rental",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.Session"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Sesi tidak aktif atau tidak ditemukan",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.Console": {
            "type": "object",
            "properties": {
                "consoleType": {
                    "$ref": "#/definitions/entity.ConsoleType"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ipAddress": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pricePerHour": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/entity.ConsoleStatus"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "entity.ConsoleStatus": {
            "type": "string",
            "enum": [
                "available",
                "in_use",
                "maintenance"
            ],
            "x-enum-comments": {
                "ConsoleStatusAvailable": "siap disewa / TV mati",
                "ConsoleStatusInUse": "sedang aktif / TV menyala",
                "ConsoleStatusMaintenance": "sedang diperbaiki"
            },
            "x-enum-varnames": [
                "ConsoleStatusAvailable",
                "ConsoleStatusInUse",
                "ConsoleStatusMaintenance"
            ]
        },
        "entity.ConsoleType": {
            "type": "string",
            "enum": [
                "PS3",
                "PS4",
                "PS5",
                "AndroidTV"
            ],
            "x-enum-varnames": [
                "ConsoleTypePS3",
                "ConsoleTypePS4",
                "ConsoleTypePS5",
                "ConsoleTypeAndroidTV"
            ]
        },
        "entity.Customer": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isMember": {
                    "description": "pelanggan member mendapat diskon otomatis",
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "entity.DiscountType": {
            "type": "string",
            "enum": [
                "percentage",
                "fixed_amount"
            ],
            "x-enum-comments": {
                "DiscountTypeFixedAmount": "diskon nominal tetap",
                "DiscountTypePercentage": "diskon persen dari total"
            },
            "x-enum-varnames": [
                "DiscountTypePercentage",
                "DiscountTypeFixedAmount"
            ]
        },
...
        "entity.Payment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "autoDiscountAmount": {
                    "description": "diskon otomatis (happy hour, member, dll)",
                    "type": "number"
                },
                "cashReceived": {
                    "description": "uang yang diterima",
                    "type": "number"
                },
                "changeAmount": {
                    "description": "kembalian",
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "discountAmount": {
                    "description": "nominal diskon yang diberikan",
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "paidAt": {
                    "type": "string"
                },
                "paymentMethod": {
                    "$ref": "#/definitions/entity.PaymentMethod"
                },
                "paymentStatus": {
                    "$ref": "#/definitions/entity.PaymentStatus"
                },
                "session": {
                    "description": "Relasi",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Session"
                        }
                    ]
                },
                "sessionId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "voucher": {
                    "$ref": "#/definitions/entity.Voucher"
                },
                "voucherId": {
                    "type": "string"
                }
            }
        },
        "entity.PaymentMethod": {
            "type": "string",
            "enum": [
                "cash"
            ],
            "x-enum-varnames": [
                "PaymentMethodCash"
            ]
        },
        "entity.PaymentStatus": {
            "type": "string",
            "enum": [
                "pending",
                "paid",
                "refunded"
            ],
            "x-enum-varnames": [
                "PaymentStatusPending",
                "PaymentStatusPaid",
                "PaymentStatusRefunded"
            ]
        },
...
        "entity.Session": {
            "type": "object",
            "properties": {
                "bookedDurationMinutes": {
                    "description": "BookedDurationMinutes adalah durasi yang dipesan di awal (contoh: 120 menit = 2 jam)",
                    "type": "integer"
                },
                "console": {
                    "description": "Relasi (join data)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Console"
                        }
                    ]
                },
                "consoleId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "customer": {
                    "$ref": "#/definitions/entity.Customer"
                },
                "customerId": {
                    "description": "nullable, walk-in",
                    "type": "string"
                },
                "durationMinutes": {
                    "type": "integer"
                },
                "endScheduledAt": {
                    "description": "EndScheduledAt adalah waktu selesai yang direncanakan (StartTime + BookedDurationMinutes)",
                    "type": "string"
                },
                "endTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/entity.SessionStatus"
                },
                "totalPrice": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "entity.SessionStatus": {
            "type": "string",
            "enum": [
                "active",
                "completed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "SessionStatusActive",
                "SessionStatusCompleted",
                "SessionStatusCancelled"
            ]
        },
...
        "entity.Voucher": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "discountType": {
                    "$ref": "#/definitions/entity.DiscountType"
                },
                "discountValue": {
                    "type": "number"
                },
                "expiresAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "maxDiscount": {
                    "description": "batas maks diskon persen (0 = tidak terbatas)",
                    "type": "number"
                },
                "maxUsage": {
                    "description": "batas total pemakaian (0 = tidak terbatas)",
                    "type": "integer"
                },
                "minPurchase": {
                    "description": "minimal total sebelum voucher berlaku",
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "usageCount": {
                    "type": "integer"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "usecase.StartSessionRequest": {
            "type": "object",
            "required": [
                "bookedDurationMinutes",
                "cashReceived",
                "consoleId"
            ],
            "properties": {
                "bookedDurationMinutes": {
                    "description": "Durasi yang dipesan dalam menit. Harus kelipatan 60 (per jam). Contoh: 60, 120, 180",
                    "type": "integer",
                    "minimum": 60,
                    "example": 120
                },
                "cashReceived": {
                    "description": "Uang tunai yang diberikan pelanggan di depan (harus \u003e= harga setelah diskon)",
                    "type": "number",
                    "example": 25000
                },
                "consoleId": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "customerId": {
                    "description": "opsional, walk-in tidak perlu",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440001"
                },
                "notes": {
                    "type": "string",
                    "example": "Rental 2 jam"
                },
                "voucherCode": {
                    "description": "Kode voucher diskon (opsional)",
                    "type": "string",
                    "example": "DISKON10"
                }
            }
        },
        "usecase.StartSessionResponse": {
            "type": "object",
            "properties": {
                "payment": {
                    "$ref": "#/definitions/entity.Payment"
                },
                "session": {
                    "$ref": "#/definitions/entity.Session"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token JWT dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}